Move webhook handling to the API endpoints.
