+++ src/HttpController/Web/JellyfinController.php
@@ -66,24 +66,6 @@ class JellyfinController
         return Response::createOk();
     }
 
-    public function handleJellyfinWebhook(Request $request) : Response
-    {
-        $webhookId = $request->getRouteParameters()['id'];
-
-        $userId = $this->userApi->findUserIdByJellyfinWebhookId($webhookId);
-        if ($userId === null) {
-            return Response::createNotFound();
-        }
-
-        $requestPayload = $request->getBody();
-
-        $this->logger->debug('Jellyfin: Webhook triggered with payload: ' . $requestPayload);
-
-        $this->jellyfinScrobbler->processJellyfinWebhook($userId, Json::decode($requestPayload));
-
-        return Response::createOk();
-    }
-
     public function regenerateJellyfinWebhookUrl() : Response
     {
         $webhookId = $this->userApi->regenerateJellyfinWebhookId($this->authenticationService->getCurrentUserId());
