+++ src/Service/WebhookUrlBuilder.php
@@ -31,6 +31,6 @@ class WebhookUrlBuilder
             return null;
         }
 
-        return rtrim($applicationUrl, '/') . '/' . $webhookType . '/' . $webhookId;
+        return rtrim($applicationUrl, '/') . '/api/' . $webhookType . '/' . $webhookId;
     }
 }
